Fix #13: Don't modify datetime format on output
Current code would modify the datetime format on the output file adding
000 to the microseconds count, so an input of "2016-03-07 12:15:27.805"
would become "2016-03-07 12:15:27.805000".

This patch changes the way we output the datetime, and instead of using
the datetime object used for sorting we will output the original
datetime string.

This will probably be useful later when we support custom log formats or
other log sources.