Add Support /var/log/messages type of logs
This patch adds support for log files that use the same datetime format
as /var/log/messages, which is in the for of Oct 15 14:11:19.

Year will from last modification to the file will be used to complete
missing year.

In the process of refactoring the read of the logs the OpenStack log
reading has been simplified to not require the presence of PID or LEVEL,
thus providing more compatibility with custom log formats (as long as
default datetime format is being used).

With this refactoring it is now easier to add new datetime formats to
the script.

Option added is --msg-logs or -ml and it also supports ALIAS and globs,
and is also affected by base directory and postfix options.

TODO: If year of file creation and file last modification are different
we should start with the cration year and then change to the next year
once the months go back.