Fix #21: KeyError with auto alias level >= 2
This patch fixes an error with auto alias generation levels 2 and 3 were
we would get a KeyError exception on certain cases.

For example with `os-log-merger -a2 2/cinder/vol.log 2/vol.log` we would
be getting KeyError: 'cinder'.

The reason for this was that we were overwriting subtrees so when we
added 2/vol.log we would be destroying the subtree generated when we
added 2/cinder/vol.log.