Fix #19: Empty output with -m option
When one of the log files is empty and we use `-m` option we will get no
output, unlike normal operation that would output contents of all other
files correctly.

This patch fixes this by removing from the files to read empty ones like
we do once we start processing the files.