Stop setting version manually
One of the reasons that pbr exists is to have versions be driven by
signed git tags instead of by content in the source repository. This has
the benefit of being both easier and preventing people from forgetting
to push a git tag when they cut a release.

Additionally, this code is doesn't work for sdist jobs in Zuul upstream
because it requires the project to be importable to run setup.py, but
sdist/wheel generation does not otherwise need to install dependencies.

Remove the override of version in setup.py and replace the hard-coded
version string in oslogmerger.py with the appropriate pbr code.

Change-Id: Ifdbb649b02d7e2fde6e8682e138c81817543fb08